backend/requirements.txt
Django==1.9.2
djangorestframework_jwt==1.6.0
djangorestframework==3.5.2
